refactor(cli): extract export command and helpers to bin/commands/export.js
Move the `export` command along with EXPORT_MARKER and its six helpers
(writeExportMarker, isExportDirectory, sanitizeFilename, uniquePathFor,
writeStream, sanitizeTitle, exportRecursive) into a dedicated module.
The helpers are re-exposed from the new module for unit testing; the
`_test` block in bin/confluence.js drops them since they no longer live
there, and tests/export.test.js imports directly from the new module.
